Move retrying to alleycats-retry
