+++ modules/alleycats/shared/src/main/scala/retry/alleycats/syntax.scala
@@ -1,0 +1,19 @@
+package retry
+package alleycats
+
+import cats.{Id, Monad}
+
+object syntax {
+  def retrying[A](
+      policy: RetryPolicy[Id],
+      wasSuccessful: A => Boolean,
+      onFailure: (A, RetryDetails) => Unit
+  )(
+      action: => A
+  )(
+      implicit
+      M: Monad[Id],
+      S: Sleep[Id]
+  ): A =
+    retryingM[A][Id](policy, wasSuccessful, onFailure)(action)
+}
